+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Noko Changelog
 
+## [Entries] - 2024-10-26
+
+- Add the ability to filter entries by day (yesterday, today, and tomorrow).
+- Add approved by type.
+
 ## [Entries] - 2024-10-05
 
 - Add the ability to list entries.